test: moved compression tests into top level module
